Authentication and security

+++ SleekChat/Api/Controllers/UsersController.cs
@@ -1,27 +1,37 @@
 ﻿using System;
 using System.Collections.Generic;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.Extensions.Options;
 using SleekChat.Core.Entities;
 using SleekChat.Data.Contracts;
 using SleekChat.Data.Helpers;
 
 namespace SleekChat.Api.Controllers
 {
+    [Authorize]
     [Route("api/[controller]")]
     [ApiController]
     public class UsersController : ControllerBase
     {
+        private readonly ICurrentUser currentUser;
         private readonly IUserData userData;
         private readonly ValidationHelper validator;
         private readonly FormatHelper formatter;
+        private readonly HttpHelper httpHelper;
         private KeyValuePair<bool, string> validationResult;
+        private readonly IOptions<AppSettings> config;
 
-        public UsersController(IUserData userData)
+        public UsersController(IUserData userData, IOptions<AppSettings> config, ICurrentUser currentUser)
         {
+            this.currentUser = currentUser;
             this.userData = userData;
+            this.config = config;
             validator = new ValidationHelper();
             formatter = new FormatHelper();
+            httpHelper = new HttpHelper();
         }
+
 
         // GET: api/users
         [HttpGet]
@@ -30,17 +40,18 @@ namespace SleekChat.Api.Controllers
             return Ok(formatter.Render(userData.GetAllUsers(), "Users", Operation.Retrieved));
         }
 
+
         // GET: api/users/id
         [HttpGet("{id}")]
         public ActionResult Get(string id)
         {
             validationResult = validator.IsBlank("User Id", id);
-            if (validationResult.Key == false) 
-                return BadRequest(formatter.Render(validationResult)); 
+            if (validationResult.Key == false)
+                return BadRequest(formatter.Render(validationResult));
 
             validationResult = validator.IsValidGuid("User Id", id);
             if (validationResult.Key == false)
-                return BadRequest(formatter.Render(validationResult)); 
+                return BadRequest(formatter.Render(validationResult));
 
             User user = userData.GetUserById(Guid.Parse(id));
             if (user == null)
@@ -49,17 +60,19 @@ namespace SleekChat.Api.Controllers
             return Ok(formatter.Render(user, "User", Operation.Retrieved));
         }
 
-        // POST: api/users
-        [HttpPost]
-        public ActionResult Post([FromBody] RequestBody reqBody)
+
+        // POST: api/users/register
+        [AllowAnonymous]
+        [HttpPost("register")]
+        public ActionResult Register([FromBody] UserReqBody reqBody)
         {
-            (string username, string email, string password, string cPassword, _) = reqBody;
+            (string username, string email, string password, string cPassword) = reqBody;
 
             validationResult = validator.IsBlank("username", username);
             if (validationResult.Key == false)
                 return BadRequest(formatter.Render(validationResult));
 
-            validationResult = validator.IsBlank("email", email); 
+            validationResult = validator.IsBlank("email", email);
             if (validationResult.Key == false)
                 return BadRequest(formatter.Render(validationResult));
 
@@ -67,9 +80,9 @@ namespace SleekChat.Api.Controllers
             if (validationResult.Key == false)
                 return BadRequest(formatter.Render(validationResult));
 
-            validationResult = validator.PasswordsMatch(password, cPassword); 
+            validationResult = validator.PasswordsMatch(password, cPassword);
             if (validationResult.Key == false)
-                return BadRequest(formatter.Render(validationResult)); 
+                return BadRequest(formatter.Render(validationResult));
 
             if (userData.UsernameAlreadyTaken(username, out _))
                 return Conflict(formatter.Render(validator.Result($"The username '{username}' is already taken.")));
@@ -77,15 +90,37 @@ namespace SleekChat.Api.Controllers
             if (userData.EmailAlreadyTaken(email, out _))
                 return Conflict(formatter.Render(validator.Result($"The email address '{email}' is already in use.")));
 
-            User newUser = userData.CreateNewUser(username, email, cPassword, true);
+            User newUser = userData.CreateNewUser(username, email, password, true);
             return Created("", formatter.Render(newUser, "User", Operation.Registered));
         }
 
+
+        // POST: api/users/authenticate
+        [AllowAnonymous]
+        [HttpPost("authenticate")]
+        public ActionResult Authenticate([FromBody] AuthReqBody reqBody)
+        {
+            validationResult = validator.IsBlank("username", reqBody.Username);
+            if (validationResult.Key == false)
+                return BadRequest(formatter.Render(validationResult));
+
+            validationResult = validator.IsBlank("password", reqBody.Password);
+            if (validationResult.Key == false)
+                return BadRequest(formatter.Render(validationResult));
+
+            AuthenticatedUser user = userData.Authenticate(reqBody, config);
+            if (user == null)
+                return BadRequest(formatter.Render(validator.Result("Username or password is invalid.")));
+
+            return Ok(formatter.Render(user, "AuthenticatedUser", Operation.Authenticated));
+        }
+
+
         // PUT: api/users/id
         [HttpPut("{id}")]
-        public ActionResult Put([FromRoute] string id, [FromBody] RequestBody reqBody)
+        public ActionResult Put([FromRoute] string id, [FromBody] UserReqBody reqBody)
         {
-            (string username, string email, string password, string cPassword, _) = reqBody;
+            (string username, string email, string password, string cPassword) = reqBody;
 
             validationResult = validator.IsBlank("User Id", id);
             if (validationResult.Key == false)
@@ -98,6 +133,13 @@ namespace SleekChat.Api.Controllers
             User user = userData.GetUserById(Guid.Parse(id));
             if (user == null)
                 return NotFound(formatter.Render(validator.Result("No such user exists.")));
+
+            if (currentUser.GetUserId() != Guid.Parse(id))
+            {
+                formatter.RenderJson(validator.Result("Sorry, you cannot edit another user's profile."), out string responseTxt);
+                httpHelper.Forbid(Response, responseTxt);
+                return null;
+            }
 
             validationResult = validator.IsBlank("username", username);
             if (validationResult.Key == false)
@@ -126,6 +168,7 @@ namespace SleekChat.Api.Controllers
             return Ok(formatter.Render(updatedUser, "User", Operation.Updated));
         }
 
+
         // DELETE: api/users/id
         [HttpDelete("{id}")]
         public ActionResult Delete([FromRoute] string id)
@@ -140,6 +183,13 @@ namespace SleekChat.Api.Controllers
 
             if (userData.GetUserById(Guid.Parse(id)) == null)
                 return NotFound(formatter.Render(validator.Result("No such user exists.")));
+
+            if (currentUser.GetUserId() != Guid.Parse(id))
+            {
+                formatter.RenderJson(validator.Result("Sorry, you cannot delete another user's profile."), out string responseTxt);
+                httpHelper.Forbid(Response, responseTxt);
+                return null;
+            }
 
             userData.DeleteUser(Guid.Parse(id));
             return Ok(formatter.Render(null, "User", Operation.Deleted));
